Revert "Move `FileCache` to it's own file and make public"
This reverts commit d6142d43e96171b8ee0c8539e38a4544abe120e9.
